refactor backend setup and add settings
