+++ .env.example
@@ -1,4 +1,4 @@
-DATABASE_URL=
-JWT_SECRET=
-ALLOWED_ORIGINS=*
-PORT=8000
+ENV=development
+DATABASE_URL=postgresql+psycopg://user:password@localhost:5432/dbname
+ALLOWED_ORIGINS=http://localhost:3000,http://127.0.0.1:3000
+JWT_SECRET=changeme
