+++ backend/requirements.txt
@@ -7,16 +7,7 @@ python-jose[cryptography]==3.3.0
 python-dotenv==1.0.1
 httpx==0.27.2
 email-validator==2.2.0
-httpx==0.27.0
-pytest==8.2.2
-email-validator==2.2.0
-
 alembic==1.13.1
-email-validator==2.2.0
 prometheus_client==0.20.0
-
-psycopg2-binary==2.9.10
-
 numpy==1.26.4
 locust==2.24.0
-
