+++ backend/settings.py
@@ -1,0 +1,15 @@
+from pydantic import BaseModel
+from dotenv import load_dotenv
+import os
+
+load_dotenv()
+
+class Settings(BaseModel):
+    ENV: str = os.getenv("ENV", "development")
+    DATABASE_URL: str | None = os.getenv("DATABASE_URL")
+    ALLOWED_ORIGINS: list[str] = [
+        o.strip() for o in os.getenv("ALLOWED_ORIGINS", "http://localhost:3000,http://127.0.0.1:3000").split(",")
+        if o.strip()
+    ]
+
+settings = Settings()
